refactor(gui): extract paths tests to sibling file
primer-gui/src/paths.rs carried a single inline trailing test module
(#[cfg(test)] mod tests) of native-lib-dir / QAIRT-path resolution
assertions. Move the whole module into a sibling src/paths/tests.rs so
the parent drops from 636 to 406 lines.

The test module is a descendant, so `use super::*` resolves to `paths`
verbatim — no visibility edits, zero churn. All 205 primer-gui lib
tests preserved; workspace 51 ok blocks, 0 failures.
